chore(txs): remove utils file
